+++ projects/angular-split/src/lib/module.ts
@@ -1,10 +1,8 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { EVENT_MANAGER_PLUGINS } from '@angular/platform-browser';
 
 import { SplitComponent } from './component/split.component';
 import { SplitAreaDirective } from './directive/splitArea.directive';
-import { UndetectedEventPlugin } from "./service/UndetectedEventPlugin";
 
 @NgModule({
     imports: [
@@ -24,11 +22,7 @@ export class AngularSplitModule {
     public static forRoot(): ModuleWithProviders {
         return {
             ngModule: AngularSplitModule,
-            providers: [{
-                provide: EVENT_MANAGER_PLUGINS,
-                useClass: UndetectedEventPlugin,
-                multi: true
-            }]
+            providers: []
         };
     }