+++ src_app/app/component/doc/doc.route.component.ts
@@ -33,19 +33,21 @@ export class DocComponent {
 
     readonly splitDoc = {
         inputs: [
-            {name: 'direction',       type: 'string',   default: '"horizontal"',  details: 'Select split direction: <code>"horizontal"</code> or <code>"vertical"</code>.'},
-            {name: 'unit',            type: 'string',   default: '"percent"',     details: `Selected unit you want to use: <code>"percent"</code> or <code>"pixel"</code>.`},
-            {name: 'gutterSize',      type: 'number',   default: '11',            details: `Gutters's size (dragging elements) in pixels.`},
-            {name: 'gutterStep',      type: 'number',   default: '1',             details: `Gutter step while moving in pixels.`},
-            {name: 'restrictMove',    type: 'boolean',  default: 'false',         details: 'Set to <code>true</code> if you want to limit gutter move to adjacent areas.'},
-            {name: 'disabled',        type: 'boolean',  default: 'false',         details: 'Disable the dragging feature (remove cursor/image on gutters). <code>(gutterClick)</code> still emits.'},
-            {name: 'dir',             type: 'string',   default: '"ltr"',         details: 'Indicates the directionality of the areas: <code>"ltr"</code> (left to right) or <code>"rtl"</code> (right to left).'},
-            {name: 'useTransition',   type: 'boolean',  default: 'false',         details: 'Add transition when toggling visibility using <code>[visible]</code> or <code>[size]</code>.<br><u>Warning: Transitions are not working for <a href="https://github.com/philipwalton/flexbugs#flexbug-16">IE/Edge/Safari</a></u>'},
+            {name: 'direction',                 type: 'string',   default: '"horizontal"',  details: 'Select split direction: <code>"horizontal"</code> or <code>"vertical"</code>.'},
+            {name: 'unit',                      type: 'string',   default: '"percent"',     details: `Selected unit you want to use: <code>"percent"</code> or <code>"pixel"</code>.`},
+            {name: 'gutterSize',                type: 'number',   default: '11',            details: `Gutters's size (dragging elements) in pixels.`},
+            {name: 'gutterStep',                type: 'number',   default: '1',             details: `Gutter step while moving in pixels.`},
+            {name: 'restrictMove',              type: 'boolean',  default: 'false',         details: 'Set to <code>true</code> if you want to limit gutter move to adjacent areas.'},
+            {name: 'disabled',                  type: 'boolean',  default: 'false',         details: 'Disable the dragging feature (remove cursor/image on gutters). <code>(gutterClick)</code> still emits.'},
+            {name: 'dir',                       type: 'string',   default: '"ltr"',         details: 'Indicates the directionality of the areas: <code>"ltr"</code> (left to right) or <code>"rtl"</code> (right to left).'},
+            {name: 'useTransition',             type: 'boolean',  default: 'false',         details: 'Add transition when toggling visibility using <code>[visible]</code> or <code>[size]</code>.<br><u>Warning: Transitions are not working for <a href="https://github.com/philipwalton/flexbugs#flexbug-16">IE/Edge/Safari</a></u>'},
+            {name: 'gutterDblClickDuration',    type: 'number',   default: '0',             details: `Milliseconds to detect a double click on a gutter. Set it around 300-500ms if you want to use <code>(gutterDblClick)</code> event.`},
         ],
         outputs: [
             {name: 'dragStart',         value: '{gutterNum: number, sizes: Array<number>}', details: 'Emit when drag starts.'},
             {name: 'dragEnd',           value: '{gutterNum: number, sizes: Array<number>}', details: 'Emit when drag ends.'},
-            {name: 'gutterClick',       value: '{gutterNum: number, sizes: Array<number>}', details: 'Emit when user clicks on a gutter.'},
+            {name: 'gutterClick',       value: '{gutterNum: number, sizes: Array<number>}', details: 'Emit when user clicks on a gutter. See <code>[gutterDblClickDuration]</code> input.'},
+            {name: 'gutterDblClick',    value: '{gutterNum: number, sizes: Array<number>}', details: 'Emit when user double clicks on a gutter. See <code>[gutterDblClickDuration]</code> input.'},
             {name: 'transitionEnd',     value: 'Array<number>',                             details: 'Emit when transition ends (could be triggered from <code>[visible]</code> or <code>[size]</code> changes).<br>Only if <code>[useTransition]="true"</code>.<br><u>Warning: Transitions are not working for <a href="https://github.com/philipwalton/flexbugs#flexbug-16">IE/Edge/Safari</a></u>'},
         ],
         class: [