+++ src_app/app/component/examples/nested.route.component.ts
@@ -14,7 +14,7 @@ import { AComponent } from './AComponent';
         <div class="container">
             <sp-example-title [type]="exampleEnum.NESTED"></sp-example-title>
             <div class="split-example" style="height: 400px;">
-                <as-split direction="horizontal" restrictMove="true" [useTransition]="true" >
+                <as-split direction="horizontal" restrictMove="true" [useTransition]="true">
                     <as-split-area size="40">
                         <as-split direction="vertical" restrictMove="true">
                             <as-split-area>