+++ src_app/app/component/examples/simple.route.component.ts
@@ -1,4 +1,5 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core';
+import { Component, ChangeDetectionStrategy, ViewChild } from '@angular/core';
+import { SplitComponent, SplitAreaDirective } from 'angular-split';
 
 import { AComponent } from './AComponent';
 
@@ -21,13 +22,28 @@ import { AComponent } from './AComponent';
         {{ testChangeDetectorRun() }}
         <div class="container">
             <sp-example-title [type]="exampleEnum.SIMPLE"></sp-example-title>
+            <h5>Percent mode:</h5>
             <div class="split-example">
-                <as-split [direction]="direction">
-                    <as-split-area size="30">
-                        <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tiam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
+                <as-split unit="percent" [direction]="direction" #split="asSplit">
+                    <as-split-area size="30" #area1="asSplitArea">
+                        <h5>Size: <b>30%</b></h5><p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tiam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
                     </as-split-area>
-                    <as-split-area size="70">
-                        <p>Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eodolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?</p>
+                    <as-split-area size="70" #area2="asSplitArea">
+                        <h5>Size: <b>70%</b></h5><p>Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eodolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?</p>
+                    </as-split-area>
+                </as-split>
+            </div>
+            <h5>Pixel mode:</h5>
+            <div class="split-example">
+                <as-split unit="pixel" [direction]="direction">
+                    <as-split-area size="120">
+                        <h5>Size: <b>120px</b></h5><p>Ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
+                    </as-split-area>
+                    <as-split-area size="*">
+                        <h5>Size: <b>*</b></h5><p>Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eodolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?</p>
+                    </as-split-area>
+                    <as-split-area size="160">
+                        <h5>Size: <b>160px</b></h5><p>Eiusmod tiam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
                     </as-split-area>
                 </as-split>
             </div>
@@ -39,5 +55,19 @@ import { AComponent } from './AComponent';
 })
 export class SimpleComponent extends AComponent {
     direction: string = 'horizontal'
+
+    @ViewChild('split') split: SplitComponent;
+    @ViewChild('area1') area1: SplitAreaDirective;
+    @ViewChild('area2') area2: SplitAreaDirective;
+
+    constructor() {
+        super();
+
+        setTimeout(() => {
+            console.log('>>> split > ', this.split);
+            console.log('>>> area1 > ', this.area1);
+            console.log('>>> area2 > ', this.area2);
+        }, 1000);
+    }
 }