+++ src_app/app/data/enum.ts
@@ -3,6 +3,7 @@ export enum ExampleEnum {
     SIMPLE = 'simple',
     MINMAX = 'minmax',
     NESTED = 'nested',
+    IFRAME = 'iframe',
     TRANSITION = 'transition',
     SYNC = 'sync',
     STYLE = 'style',