+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="3.0.0-XXX"></a>
+# 3.0.0-XXX (2019-02-XX)
+
+* **Pixel mode:** Now your can choose to work with `pixel` or `percent` mode. In `pixel` mode, an area with wildcard size (`size="*"`) is mandatory.
+* **minSize & maxSize area:** Now you can set minimum and maximum sizes for each areas whatever the current mode (Only exception is the wildcard area size in `pixel` mode).
+* **gutterDblClick event:** Double click on gutter is now catchable with `(gutterDblClick)`.
+* **gutterDblClickDuration property:** Specify duration between 2 clicks to consider a double click `(gutterDblClick)` event.
+* **exportAs:** You can access `SplitComponent` as `<as-split #split="asSplit">` and `SplitAreaDirective` as `<as-split-area #area1="asSplitArea">` from template variables easily with `ViewChild`.
+* **Internal:** Remove `EventManagerPlugin` and start/end/click subscribers stuff which was added to avoid some change detection runs but adding too much complexity.. I stopped going against the framework, clearer now!
+
 <a name="2.0.1"></a>
 # 2.0.1 (2018-12-21)